Rename `api` to `boards`.
